Save image as png
Changing the way image is stored saves a lot of memory (png is
compressed contrary to ppm format). It is implemented with image crate
and while it would be better to use png crate directly (to have less
dependencies), the documentation was worse and it was faster to use
image crate.