Implement antialiasing
Implementation details: For every pixel, multiple rays is computed
(there's slight deviation of the rays within the pixel) and all of their
colors are combined into one color, that is saved into the image. How
many rays will be computed is given by `SAMPLES_PER_PIXEL` constant.

Due to the need for adding these colors and than dividing them, the
implementation of Color struct with u8 fields wasn't good enough, and
instead of editing and refactoring of this struct I created a new Color
module, that implements R, G, B fields as fractions in range <0, 1>.
This new module is imported as FColor.

The antialiasing is a significant performance hit, but it could be
easily run in parallel on multiple cores. Also SIMD could be used to
speed it up.